+++ rubikpi/__init__.py
@@ -8,6 +8,6 @@ the Free Software Foundation, either version 3 of the License, or
 (at your option) any later version.
 """
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"
 __app_name__ = "RubikPI"
 __author__ = "Gwilherm Kerherve"